+++ src/main/java/dev/nincodedo/ninbot/common/message/SlashCommandEventMessageAction.java
@@ -1,0 +1,8 @@
+package dev.nincodedo.ninbot.common.message;
+
+public class SlashCommandEventMessageAction implements MessageAction{
+    @Override
+    public void executeActions() {
+
+    }
+}
