+++ src/main/java/dev/nincodedo/ninbot/components/poll/PollUserChoiceListener.java
@@ -2,7 +2,7 @@ package dev.nincodedo.ninbot.components.poll;
 
 import dev.nincodedo.ninbot.common.Constants;
 import dev.nincodedo.ninbot.common.StatAwareListenerAdapter;
-import dev.nincodedo.ninbot.common.message.MessageAction;
+import dev.nincodedo.ninbot.common.message.MessageReceivedEventMessageAction;
 import dev.nincodedo.ninbot.components.stats.StatManager;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 
@@ -36,7 +36,7 @@ public class PollUserChoiceListener extends StatAwareListenerAdapter {
                     refMessage.editMessage(poll.build()).queue();
                     pollAnnouncementSetup.setupAnnounce(poll, event.getJDA().getShardManager(), refMessage);
                 } else {
-                    new MessageAction().setOverrideMessage(event.getMessage())
+                    new MessageReceivedEventMessageAction().setOverrideMessage(event.getMessage())
                             .addUnsuccessfulReaction()
                             .executeActions();
                 }
