+++ src/main/java/dev/nincodedo/ninbot/components/stats/StatsCommand.java
@@ -1,7 +1,7 @@
 package dev.nincodedo.ninbot.components.stats;
 
-import dev.nincodedo.ninbot.common.message.MessageAction;
 import dev.nincodedo.ninbot.common.command.AbstractCommand;
+import dev.nincodedo.ninbot.common.message.MessageReceivedEventMessageAction;
 import dev.nincodedo.ninbot.components.config.ConfigConstants;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.MessageBuilder;
@@ -32,8 +32,8 @@ public class StatsCommand extends AbstractCommand {
 
     //TODO implement SlashCommand
     @Override
-    protected MessageAction executeCommand(PrivateMessageReceivedEvent event) {
-        MessageAction messageAction = new MessageAction(event);
+    protected MessageReceivedEventMessageAction executeCommand(PrivateMessageReceivedEvent event) {
+        MessageReceivedEventMessageAction messageAction = new MessageReceivedEventMessageAction(event);
 
         return messageAction;
     }
